ripple shader

+++ src/scripts/map/map.gd
@@ -16,9 +16,10 @@ static var item_type: Dictionary[String, Item] = {
 
 var map: Dictionary[Vector2i, Block] = {}
 
-@onready var item_layer: TileMapLayer = $ItemLayer
-@onready var element_layer: TileMapLayer = $ElementLayer
-@onready var grass_layer: TileMapLayer = $GrassLayer
+@onready var item_layer: TileMapLayer = $CanvasGroup/ItemLayer
+@onready var element_layer: TileMapLayer = $CanvasGroup/ElementLayer
+@onready var grass_layer: TileMapLayer = $CanvasGroup/GrassLayer
+@onready var audio_stream_player: AudioStreamPlayer2D = $AudioStreamPlayer2D
 
 func _ready() -> void:
 	place_element(Vector2i(0,0), element_type["Sand"])
@@ -81,12 +82,17 @@ func a_block_was_updated(pos: Vector2i, block: Block):
 	item_layer.block_updated(pos, block)
 
 
+func play_sound(position: Vector2, audio: AudioStream):
+	audio_stream_player.global_position = position
+	audio_stream_player.stream = audio
+	audio_stream_player.play()
+
+
 var selected = 0
 
 func _input(event: InputEvent) -> void:
 	if event is InputEventKey and event.is_pressed() and event.keycode == Key.KEY_SPACE:
 		selected = (selected+1) % 2
-		print("selected ", selected)
 	
 	if event is InputEventMouse:
 		var pressed = (
@@ -102,7 +108,7 @@ func _input(event: InputEvent) -> void:
 			var grid_pos = item_layer.local_to_map(pos)
 			
 			if Input.is_mouse_button_pressed(MOUSE_BUTTON_LEFT):
-				if Input.is_key_pressed(KEY_SHIFT):
+				if Input.is_key_pressed(KEY_SHIFT) and OS.has_feature("debug"):
 					var pickable = Pickable.new()
 					var elem
 					if selected == 0:
@@ -111,14 +117,15 @@ func _input(event: InputEvent) -> void:
 						elem = element_type["Grass"]
 					place_element(grid_pos, elem)
 				else:
-					var game = $".."
+					var game: Game = $".."
 					var pickable = game.get_hand()
-					print(grid_pos)
 					var interact_has_worked = interact_at(grid_pos, pickable)
 					if interact_has_worked:
+						if pickable:
+							play_sound(get_global_mouse_position(), pickable.audio)
 						game.use_hand()
 					
-			elif Input.is_mouse_button_pressed(MOUSE_BUTTON_RIGHT):
+			elif Input.is_mouse_button_pressed(MOUSE_BUTTON_RIGHT) and OS.has_feature("debug"):
 				destroy_block(grid_pos)
 
 var timer = 0
@@ -127,9 +134,9 @@ func _process(delta: float) -> void:
 	timer += delta
 	if timer > 1:
 		timer -= 1
-
+		
 		for pos in map.keys():
 			var block = map[pos]
 			block._on_random_tick()
-	
+
